refact: create Loading component
