+++ src/components/Loading/Loading.module.scss
@@ -1,0 +1,11 @@
+.loading {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translateX(-50%);
+
+    &__img {
+        width: 50px;
+        height: 50px;
+    }
+}
