+++ src/components/Main/Main.module.scss
@@ -13,15 +13,3 @@
         }
     }
 } 
-
-.loading {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translateX(-50%);
-
-    &__img {
-        width: 50px;
-        height: 50px;
-    }
-}
