+++ src/components/Main/variants/index.ts
@@ -1,20 +1,3 @@
-//Main
-export const loadingVariants = {
-  hidden: {
-    opacity: 0
-  },
-  visible: {
-    y: [-30, 0],
-    opacity: 1,
-    transition: {
-      y: {
-        yoyo: Infinity,
-        duration: 0.4,
-        ease: 'easeOut'
-      }
-    }
-  }
-}
 //User
 export const itemUserVariants = {
   hidden: { y: 60, opacity: 0 },
